+++ CHANGELOG.md
@@ -4,7 +4,8 @@ TODO Next Version
 -------------------------------------------------------------------------------------------
 FIXES AND ENHANCEMENTS
  - #21: AbstractCoberturaCoverageMetric: Make logger instance field to enable verifying logged warnings.
- - #23: CyclomaticComplexityMetric: Support Groovy 3 syntax: do/while loop, elvis assignment (?=), safe index, (e.g. x?[1, 2]).
+ - #23: CyclomaticComplexityMetric: Support Groovy 3 syntax: do/while loop, elvis assignment (`?=`), safe index, (e.g. `x?[1, 2]`).
+ - #24: AbcMetric: Support Groovy 3 syntax: try-with-resources, `===`, `!==`, `!in`, `!instanceof`, elvis assignment (`?=`), safe index, (e.g. `x?[1, 2]`). Also add support for existing `in` and `instanceof` operators.
 
 INFRASTRUCTURE AND TESTS
  - #20: Requires Java 8 or later.